Use the regex we claim to be using in the README
We just use the date part of the ISO8601 period expression. The proper regex
for validating it is already in the ValidatorClass for Caluma in our README.

With this commit, we're actually using this regex to validate the
provided interval.

This commit also refactors the whole interval parsing and it's corresponding
tests.